requirements.txt
streamlit
pandas
psycopg2-binary
Werkzeug
streamlit-drawable-canvas
streamlit-folium
